C/S: Changed TTL to Server TTL Time
This is needed for the situation where this machine is still considered the server, even though it died. It would start sending out summaries before the HeartbeatSend can update the TTL to server time. If this isn't already considered the server, it will send a normal TTL heartbeat to the right server, and that summary received will overwrite this heartbeat in the cache.